The changes made in the last commit to functions linked to corridor data have been reverted. The extractStationsFromCorridor() function has been modified to read all routes within a corridor, even if they have different start addresses.
